refactor: split lex.go into buf.go (I/O) + charclass.go (predicates)
