+++ buf.go
@@ -1,0 +1,95 @@
+// Copyright 2014 The Go Authors.  All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// Buffered I/O primitives for reading raw bytes from a PDF stream.
+
+package pdf
+
+import (
+	"fmt"
+	"io"
+)
+
+// A buffer holds buffered input bytes from the PDF file.
+type buffer struct {
+	r           io.Reader // source of data
+	buf         []byte    // buffered data
+	pos         int       // read index in buf
+	offset      int64     // offset at end of buf; aka offset of next read
+	tmp         []byte    // scratch space for accumulating token
+	unread      []token   // queue of read but then unread tokens
+	allowEOF    bool
+	allowObjptr bool
+	allowStream bool
+	eof         bool
+	key         []byte
+	useAES      bool
+	objptr      objptr
+	depth       int
+}
+
+// newBuffer returns a new buffer reading from r at the given offset.
+func newBuffer(r io.Reader, offset int64) *buffer {
+	return &buffer{
+		r:           r,
+		offset:      offset,
+		buf:         make([]byte, 0, 4096),
+		allowObjptr: true,
+		allowStream: true,
+	}
+}
+
+func (b *buffer) readByte() byte {
+	if b.pos >= len(b.buf) {
+		b.reload()
+		if b.pos >= len(b.buf) {
+			return '\n'
+		}
+	}
+	c := b.buf[b.pos]
+	b.pos++
+	return c
+}
+
+func (b *buffer) errorf(format string, args ...any) {
+	panic(fmt.Errorf(format, args...))
+}
+
+func (b *buffer) reload() bool {
+	n := cap(b.buf) - int(b.offset%int64(cap(b.buf)))
+	n, err := b.r.Read(b.buf[:n])
+	if n == 0 && err != nil {
+		b.buf = b.buf[:0]
+		b.pos = 0
+		if b.allowEOF && err == io.EOF {
+			b.eof = true
+			return false
+		}
+		b.errorf("malformed PDF: reading at offset %d: %v", b.offset, err)
+		return false
+	}
+	b.offset += int64(n)
+	b.buf = b.buf[:n]
+	b.pos = 0
+	return true
+}
+
+func (b *buffer) seekForward(offset int64) {
+	for b.offset < offset {
+		if !b.reload() {
+			return
+		}
+	}
+	b.pos = len(b.buf) - int(b.offset-offset)
+}
+
+func (b *buffer) readOffset() int64 {
+	return b.offset - int64(len(b.buf)) + int64(b.pos)
+}
+
+func (b *buffer) unreadByte() {
+	if b.pos > 0 {
+		b.pos--
+	}
+}
